requirements.txt
    beautifulsoup4>=4.14.2
    bs4>=0.0.2
    einops>=0.8.1
    jupyterlab>=4.5.0
    jupyterlab-slideshow>=0.3.4
    langchain>=1.1.0
    langchain-chroma>=1.0.0
    langchain-community>=0.4.1
    langchain-huggingface>=1.1.0
    langchain-mcp-adapters>=0.2.1
    langchain-mistralai>=1.1.0
    langchain-text-splitters>=1.0.0
    sentence-transformers>=5.1.2
    tavily-python>=0.7.14
    unstructured[all-docs]>=0.18.21
    ipywidgets
    peft
    python-dotenv
    mlflow
    langgraph>=0.2.0
    fpdf2